Simplify the workflow via Makefile (#42)
This set of changes tries to simplify running things a bit and adds a few things
to the helper scripts.

The gist of the change is the "preferred workflow" is now:
  * make build
  * make rancher-up
  * make rancher-destroy

This focuses on the container model for running ez-rancher.  Those that want to
run locally with their own terraform install, "good on ya", but for those
wanting to run the "ez" version here it is.

Also removes the implicit image build.